Move GlslParser to its own package.
