+++ src/commonTest/kotlin/baaahs/gl/glsl/GlslCodeSpec.kt
@@ -1,6 +1,7 @@
 package baaahs.gl.glsl
 
 import baaahs.gl.expectValue
+import baaahs.gl.glsl.parser.GlslParser
 import baaahs.gl.override
 import baaahs.gl.undefined
 import baaahs.only
